• ProcessingMechanisms   - DNDMechanism implemented - still needs _execute
• IntegratorFunctions
  - DND implemented - needs correct referencing to metrics

• ObjectiveFunctions
  - Distance: cosine added as a metric - still needs LLVM version
