+++ psyneulink/library/components/mechanisms/processing/integrator/__init__.py
@@ -1,5 +1,8 @@
 from . import ddm
+from . import dnd
 
 from .ddm import *
+from .dnd import *
 
 __all__ = list(ddm.__all__)
+__all__.extend(dnd.__all__)
